refactor: 중복된 코드 정리 및 삭제
