+++ src/hooks/useUser.tsx
@@ -4,6 +4,7 @@ import { useContext } from "react";
 const useUser = () => {
   const { currentUser, dispatch } = useContext(UserContext);
 
+  // user 토글 함수
   const toggleUser = (name: string): void => {
     dispatch({
       type: "TOGGLE_USER",
